more simplifications + CI
