+++ src/packs.rs
@@ -1,5 +1,4 @@
 pub struct Packs {}
-
 impl Packs {
     pub fn greet() -> () {
         println!("Hello! This CLI is under construction.")
